Refactor: reorganized dir layout
