+++ reddit_data/data_models/__init__.py
@@ -1,0 +1,1 @@
+from .reddit_comment import RedditComment
